setup event bus to listen for touches.
On mobile browsers hover is replaced with touch to trigger the tooltip. This causes the tool tip to not close when you touch anywhere else on the document.

This adds an event bus to listen for touches on the document and then resets the visible value to null.

We change from false to null here, because we don’t want to override any of the user overrides. Eg; if you as a developer manually control the visibility of a tooltip then we won’t try to automatically open or close it.